Changes to PATH and MANPATH.
We use ~/Local/bin instead of ~/Bin or ~/Local/Bin.
We use ~/Local/man instead of ~/Man or ~/Local/Man.

Also, fixed a problem where we were putting $JAVA_HOME/bin in
PATH when $JAVA_HOME was undefined.  This caused /bin to be earlier
in the PATH than ~/Local/bin, which makes it impossible to install
local versions of tools like zsh.